Add ability to move forward/backward in slideshow.
At the beginning of the app's run, it maps all the image files in
the directory and keeps them in a list. The list is shuffled to
have a random, different order every time. Each time the delay time
passes, the saved index is increased to get the next image.

When the left/right keys are pressed, the `on_key_press` method is
activated and a boolean value is set to true (indicating they were
pressed), and the counter for pictures to move is decreased/increased
(left/right). When the boolean is set to True, the image changing
process is started again in `on_draw`.